+++ ZenLib.Tests/UsageTests.cs
@@ -1288,7 +1288,7 @@ namespace ZenLib.Tests
         [TestMethod]
         public void TestExactSeqs()
         {
-            var f = new ZenFunction<FSeq<int>, ushort>(l => l.Length());
+            var f = new ZenFunction<FSeq<int>, BigInteger>(l => l.Length());
             f.Find((i, o) => true);
         }
